# Faster freelist updates