--- a/page_test.go
+++ b/page_test.go
@@ -1,6 +1,7 @@
 package bolt
 
 import (
+	"reflect"
 	"testing"
 )
 
@@ -26,4 +27,19 @@ func TestPage_typ(t *testing.T) {
 // Ensure that the hexdump debugging function doesn't blow up.
 func TestPage_dump(t *testing.T) {
 	(&page{id: 256}).hexdump(16)
+}
+
+func TestPgids_Merge(t *testing.T) {
+	a := pgids{4, 5, 6, 10, 11, 12, 13, 27}
+	b := pgids{1, 3, 8, 9, 25, 30}
+	c := a.Merge(b)
+	if !reflect.DeepEqual(c, pgids{1, 3, 4, 5, 6, 8, 9, 10, 11, 12, 13, 25, 27, 30}) {
+		t.Errorf("mismatch: %v", c)
+	}
+	a = pgids{4, 5, 6, 10, 11, 12, 13, 27, 35, 36}
+	b = pgids{8, 9, 25, 30}
+	c = a.Merge(b)
+	if !reflect.DeepEqual(c, pgids{4, 5, 6, 8, 9, 10, 11, 12, 13, 25, 27, 30, 35, 36}) {
+		t.Errorf("mismatch: %v", c)
+	}
 }