Since freelist is kept sorted we can update it by merging the sorted pgid lists rather than appending and resorting the whole list. We have DBs in production reaching 10M pages in the freelist and about 20K pages in the pending lists. The benchmark is an attempt to prove that at these sizes the merge can be nearly 100x faster than the append/sort approach.

Append/Sort:
```
Martins-MacBook-Pro-5:bolt martin$ go test -v -bench=FreelistRelease -run None -short
seed: 56446
quick settings: count=5, items=1000, ksize=1024, vsize=1024
PASS
Benchmark_FreelistRelease10K	    1000	   1301767 ns/op
Benchmark_FreelistRelease100K	     100	  15634815 ns/op
Benchmark_FreelistRelease1000K	      10	 182048520 ns/op
Benchmark_FreelistRelease10000K	       1	2187250681 ns/op
ok  	github.com/boltdb/bolt	11.825s
```

Merge:
```
Martins-MacBook-Pro-5:bolt martin$ go test -v -bench=FreelistRelease -run None -short
seed: 27571
quick settings: count=5, items=1000, ksize=1024, vsize=1024
PASS
Benchmark_FreelistRelease10K	   30000	     50734 ns/op
Benchmark_FreelistRelease100K	    5000	    210540 ns/op
Benchmark_FreelistRelease1000K	     500	   3132026 ns/op
Benchmark_FreelistRelease10000K	      30	  37413055 ns/op
ok  	github.com/boltdb/bolt	19.631s
```

This isn't a theoretical issue, we have a production scenario with large freelists where freelist sorting is taking nearly 90% of CPU time. I suspect that slowness of the freelist updates could significantly contribute to explosive freelist growth as well if it extends transaction duration by seconds.

@benbjohnson, I'm going to test this branch in some of our production scenarios next, but I would love your thoughts on this change, in case I'm missing something. Thanks!

/cc @snormore