@mkobetic 2015-06-13
Sorry, had to update the benchmark and the results after realizing that I wasn't benchmarking sorted lists.