@mkobetic 2015-06-13
This is how our write times dropped with the freelist fix, the metric is the time it took to update 1000 rows in milliseconds.

![screen shot 2015-06-12 at 11 29 02 pm](https://cloud.githubusercontent.com/assets/871693/8142846/eb7f3c6a-115a-11e5-9fef-3b7e18ba7f61.png)