@benbjohnson 2015-06-13
Overall it lgtm. The merge is a little confusing and I think the binary search could be slow. I tried an alternate, simpler version where it just walks both lists:

https://github.com/benbjohnson/bolt/commit/f7026f989ad0310bb2559fb87f5747f55cb89369

It ends up being faster for large sets but slower for small sets. :-/ although i don't think the perf is big enough to make a difference in practice.

Can you run this for a little while in production and if it's working well then let's merge.