@mkobetic 2015-06-14
I'll give it a try, but the reasoning I had for using the search is that we're seeing about 400 times less pages in the pending lists than in the main free list, so assuming perfectly random pending list you're still looking at 400 element runs on the freelist on average. In my experience bulk copying primitives save a lot compared to individual element access that has to bounds check every single access call. I don't know if the bin search function is optimized or not, but I was hoping that even if it isn't it would still be fewer indexed accesses than individual copying. Anyway, I admit these are just conjectures.