@benbjohnson 2015-06-14
That's a good point about the bounds checking. Sticking with `sort.Search()` works for me.