@mkobetic 2015-06-15
Sounds good. Feel free to change anything to make it more readable, I don't mind. Otherwise, FWIW, we've been running with the original patch in production for almost 2 days now without any adverse effects so far.