@mkobetic 2015-06-15
BTW, it seemed to also speed up our query time average about 5 times (from ~100ms to ~20ms). I guess the freelist updates were eating up majority of the query time.