@benbjohnson 2015-06-16
@mkobetic I added a `testing/quick` test to validate the merge against random input and it's failing. Can you take a look at the issue:

https://github.com/benbjohnson/bolt/commit/fc360dd4581d1ef71a71332915fd5157b77288ed#diff-126a90c17ac91c89ec71177a7627dcf8R50

Here's what I'm getting:

```
--- FAIL: TestPgids_merge_quick (0.00s)
	page_test.go:68: 
		exp=[220 393 435 479 491 560 616 626 658 810 1104 1355 1358 1506 1786 1829 1898 2005 2074 2125 2407 2640 2886 3072 3073 3075 3089 3129 3343 3347 3418 3671 3853 3865 4552 5418 5507 6092 6142 6362 6375 6615 6756 6764 6844 6864 7025 7128 7766 7802 7923 8354 8398 8804 9010 9574 9775 9914 9938 9961]
		got=[6864 7923 1506 9914 9775 3865 4552 7025 1358 7128 8804 1355 9010 616 9961 3671 3343 435 6844 5418 658 2125 6362 3418 9938 479 6756 491 7802 2886 6764 3073 1829 626 3129 3347 3075 8354 810 3072 2074 8398 9574 6142 3089 220 2407 5507 3853 560 1898 6615 1104 6092 7766 1786 2640 393 6375 2005]
	page_test.go:74: #1: failed on input bolt.pgids{0x1ad0, 0x1ef3, 0x5e2, 0x26ba, 0x262f, 0xf19, 0x11c8, 0x1b71, 0x54e, 0x1bd8, 0x2264, 0x54b, 0x2332, 0x268, 0x26e9, 0xe57, 0xd0f, 0x1b3, 0x1abc, 0x152a, 0x292, 0x84d, 0x18da, 0xd5a}, bolt.pgids{0x26d2, 0x1df, 0x1a64, 0x1eb, 0x1e7a, 0xb46, 0x1a6c, 0xc01, 0x725, 0x272, 0xc39, 0xd13, 0xc03, 0x20a2, 0x32a, 0xc00, 0x81a, 0x20ce, 0x2566, 0x17fe, 0xc11, 0xdc, 0x967, 0x1583, 0xf0d, 0x230, 0x76a, 0x19d7, 0x450, 0x17cc, 0x1e56, 0x6fa, 0xa50, 0x189, 0x18e7, 0x7d5}
```