@mkobetic 2015-06-16
I need to double check what quick.Check does but, at a quick glance it seems it may be feeding the merge function with unsorted lists. Is that possible?