@benbjohnson 2015-06-16
Ah, that's it. Thanks, Martin. Fixed! I'll merged it all in. I just added some comments and changed `Merge()` to `merge()` since it doesn't need to be exported.