@benbjohnson 2015-06-16
I rebased the commits and merged into `master`: https://github.com/boltdb/bolt/commit/04a3e85793043e76d41164037d0d7f9d53eecae3