@benbjohnson 2015-06-16
Also, to answer the `quick.Check` question, it takes the `func` you pass into it and generates random data a bunch of times (`100` times by default). You can adjust it with the `-quickchecks COUNT` flag.