@mkobetic 2015-06-25
Just an FYI, after rebuilding our production DBs with the faster freelist updates, our freelist sizes went from GBs to MBs. So I'm fairly confident now that the slow updates were a significant contributor to freelist growth.